pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "asvspoof5-domain-invariant-cm"
version = "0.1.0"
description = "Domain-invariant speech deepfake detection on ASVspoof 5 (Track 1) using DANN and mechanistic analysis."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Mike Cooper", email = "mike.cooper@student.uva.nl" }]
keywords = [
    "audio",
    "deepfake",
    "asvspoof",
    "speech",
    "domain-adaptation",
    "domain-adversarial",
    "dann",
    "wavlm",
    "wav2vec2",
    "interpretability",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=2.0",
    "pandas>=2.2",
    "scipy>=1.14",
    "scikit-learn>=1.5",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "soundfile>=0.12",
    "librosa>=0.10.2",
    "numba>=0.60",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "torch>=2.5",
    "torchaudio>=2.5",
    "transformers>=4.45",
    "datasets>=3.0",
    "accelerate>=1.0",
    "einops>=0.8",
    "rich>=13.9",
    "wandb>=0.18",
    "notebook>=7.5.3",
    "jupyterlab>=4.5.4",
    "umap-learn",
]

[project.optional-dependencies]
notebooks = [
    "jupyter>=1.0",
    "notebook>=7.0",
    "ipykernel>=6.0",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "pandas>=2.2",
    "plotly>=5.0",
    "kaleido>=0.2",
    "tabulate>=0.9",
]
dev = [
    "asvspoof5-domain-invariant-cm[notebooks]",
    "ruff>=0.8",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pre-commit>=4.0",
]
xai = [
    "captum>=0.7",
]
trillsson = [
    "tensorflow",
    "tensorflow-hub",
    "pyarrow>=14.0",
    "soundfile>=0.12",
    "scipy>=1.11",
]

[tool.hatch.build.targets.wheel]
packages = ["src/asvspoof5_domain_invariant_cm"]

[tool.ruff]
line-length = 80
target-version = "py310"
extend-select = ["I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
    "slow: marks tests as slow (downloads models, requires network)",
]
